Actually set up cross-cluster replication, kinda
